livekit-tests/requirements.txt
python-dotenv
livekit-agents
livekit-agents[silero,neuphonic]
aiohttp
pytest
pytest-asyncio